Test : Post, Comment, Board 레포지토리 및 엔티티 관계 테스트 성공

+++ src/main/kotlin/soma/achoom/zigg/comment/entity/Comment.kt
@@ -16,7 +16,7 @@ class Comment(
 
     var textComment: String,
 
-    var likes: Long = 0,
+    var likes : Int = 0,
 
     @OneToMany(cascade = [CascadeType.PERSIST], orphanRemoval = true)
     val replies: MutableList<Comment> = mutableListOf(),
